+++ tests/Builder/CreateStatementTest.php
@@ -57,6 +57,23 @@ class CreateStatementTest extends TestCase
         );
     }
 
+    public function testBuilderCollate()
+    {
+        $parser = new Parser(
+            'CREATE TABLE IF NOT EXISTS t1 (' .
+            " c1 varchar(11) NOT NULL DEFAULT '0' COLLATE 'utf8_czech_ci' COMMENT 'xxx'" .
+            ') ENGINE=MyISAM'
+        );
+        $stmt = $parser->statements[0];
+
+        $this->assertEquals(
+            "CREATE TABLE IF NOT EXISTS t1 (\n" .
+            "  `c1` varchar(11) NOT NULL DEFAULT '0' COLLATE 'utf8_czech_ci' COMMENT 'xxx'\n" .
+            ') ENGINE=MyISAM',
+            $stmt->build()
+        );
+    }
+
     public function testBuilderDefaultComment()
     {
         $parser = new Parser(